# internal/repositories/employee/queries/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.25.0

package employee

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Employee struct {
	ID           int32       `json:"id"`
	Name         pgtype.Text `json:"name"`
	Surname      pgtype.Text `json:"surname"`
	Phone        pgtype.Text `json:"phone"`
	CompanyID    pgtype.Int4 `json:"company_id"`
	PassportID   pgtype.Int4 `json:"passport_id"`
	DepartmentID pgtype.Int4 `json:"department_id"`
}